create a utils.go
