+++ pkg/core/engine/main.go
@@ -2,17 +2,11 @@ package engine
 
 import (
 	"errors"
-	"os"
-
-	"github.com/sirupsen/logrus"
 
 	"github.com/updatecli/updatecli/pkg/core/config"
 	"github.com/updatecli/updatecli/pkg/core/pipeline"
 	"github.com/updatecli/updatecli/pkg/core/reports"
-	"github.com/updatecli/updatecli/pkg/core/result"
 	"github.com/updatecli/updatecli/pkg/core/tmp"
-
-	"path/filepath"
 )
 
 var (
@@ -32,58 +26,4 @@ type Engine struct {
 func (e *Engine) Clean() (err error) {
 	err = tmp.Clean()
 	return
-}
-
-// GetFiles return an array with every valid files.
-func GetFiles(root []string) (files []string) {
-	if len(root) == 0 {
-		// Updatecli tries to load the file updatecli.yaml if no manifest provided
-		// If updatecli.yaml doesn't exists then Updatecli parses the directory updatecli.d for any manifests.
-		// if there is no manifests in the directory updatecli.d then Updatecli returns no manifest files.
-		_, err := os.Stat(config.DefaultConfigFilename)
-		if !errors.Is(err, os.ErrNotExist) {
-			logrus.Debugf("Default Updatecli manifest detected %q", config.DefaultConfigFilename)
-			return []string{config.DefaultConfigFilename}
-		}
-
-		fs, err := os.Stat(config.DefaultConfigDirname)
-		if errors.Is(err, os.ErrNotExist) {
-			return []string{}
-		}
-
-		if fs.IsDir() {
-			logrus.Debugf("Default Updatecli manifest directory detected %q", config.DefaultConfigDirname)
-			root = []string{config.DefaultConfigDirname}
-		}
-	}
-
-	for _, r := range root {
-		err := filepath.Walk(r, func(path string, info os.FileInfo, err error) error {
-			if err != nil {
-				logrus.Errorf("\n%s File %s: %s\n", result.FAILURE, path, err)
-				os.Exit(1)
-			}
-			if info.Mode().IsRegular() {
-				files = append(files, path)
-			}
-			return nil
-		})
-
-		if err != nil {
-			logrus.Errorf("err - %s", err)
-		}
-	}
-
-	// Remove duplicates manifest files
-	result := []string{}
-	exist := map[string]bool{}
-
-	for v := range files {
-		if !exist[files[v]] {
-			exist[files[v]] = true
-			result = append(result, files[v])
-		}
-	}
-
-	return result
 }
